templates/docker/python/requirements.txt
# Base requirements for Python Docker container
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-exporter-gcp-trace
google-cloud-trace
google-cloud-logging
